Moved Intel intrinsics generator to separate file
