+++ src/midend/programTransformation/CMakeLists.txt
@@ -44,6 +44,7 @@ add_library(midend_pt OBJECT
   astOutlining/GenerateCall.cc
   ompLowering/omp_lowering.cpp
   ompLowering/omp_simd.cpp
+  ompLowering/intel_simd.cpp
   astInlining/isPotentiallyModified.C
   astInlining/replaceExpressionWithStatement.C
   astInlining/inliner.C
